app/requirements.txt
fastapi
uvicorn
pydantic
passlib[bcrypt]
python-jose
databases
asyncpg
httpx
sqlalchemy
psycopg2-binary
python-dotenv

